+++ leptos_macro/Cargo.toml
@@ -2,7 +2,6 @@
 name = "leptos_macro"
 version = { workspace = true }
 authors = ["Greg Johnston"]
-license = "MIT"
 repository = "https://github.com/leptos-rs/leptos"
 description = "view macro for the Leptos web framework."
 readme = "../README.md"
@@ -13,90 +12,24 @@ edition.workspace = true
 proc-macro = true
 
 [dependencies]
-attribute-derive = { version = "0.10.3", features = ["syn-full"] }
-cfg-if = "1.0"
-html-escape = "0.2.13"
-itertools = { workspace = true }
-prettyplease = "0.2.25"
-proc-macro-error2 = { version = "2.0", default-features = false }
-proc-macro2 = "1.0"
-quote = "1.0"
-syn = { version = "2.0", features = ["full"] }
-rstml = "0.12.0"
-leptos_hot_reload = { workspace = true }
-server_fn_macro = { workspace = true }
-convert_case = { workspace = true }
-uuid = { version = "1.11", features = ["v4"] }
-tracing = { version = "0.1.41", optional = true }
-
-[dev-dependencies]
-log = "0.4.22"
-typed-builder = "0.20.0"
-trybuild = { workspace = true }
-leptos = { path = "../leptos" }
-leptos_router = { path = "../router", features = ["ssr"] }
-server_fn = { path = "../server_fn", features = ["cbor"] }
-insta = "1.41"
-serde = "1.0"
-
-[build-dependencies]
-rustc_version = "0.4.1"
+leptos_macro_core.workspace = true
 
 [features]
-csr = []
-hydrate = []
-ssr = ["server_fn_macro/ssr", "leptos/ssr"]
-nightly = ["server_fn_macro/nightly"]
-tracing = ["dep:tracing"]
-islands = []
-trace-components = []
-trace-component-props = []
-actix = ["server_fn_macro/actix"]
-axum = ["server_fn_macro/axum"]
-generic = ["server_fn_macro/generic"]
+csr = ["leptos_macro_core/csr"]
+hydrate = ["leptos_macro_core/hydrate"]
+ssr = ["leptos_macro_core/ssr"]
+nightly = ["leptos_macro_core/nightly"]
+tracing = ["leptos_macro_core/tracing"]
+islands = ["leptos_macro_core/islands"]
+trace-components = ["leptos_macro_core/trace-components"]
+trace-component-props = ["leptos_macro_core/trace-component-props"]
+actix = ["leptos_macro_core/actix"]
+axum = ["leptos_macro_core/axum"]
+generic = ["leptos_macro_core/generic"]
 # Having an erasure feature rather than normal --cfg erase_components for the proc macro crate is a workaround for this rust issue:
 # https://github.com/rust-lang/cargo/issues/4423
 # TLDR proc macros will ignore RUSTFLAGS when --target is specified on the cargo command.
 # This works around the issue by the non proc-macro crate which does see RUSTFLAGS enabling the replacement feature on the proc-macro crate, which wouldn't.
 # This is automatic as long as the leptos crate is depended upon,
 # downstream usage should never manually enable this feature.
-__internal_erase_components = []
-
-[package.metadata.cargo-all-features]
-denylist = ["tracing", "trace-component-props", "trace-components"]
-skip_feature_sets = [
-  [
-    "csr",
-    "hydrate",
-  ],
-  [
-    "hydrate",
-    "csr",
-  ],
-  [
-    "hydrate",
-    "ssr",
-  ],
-  [
-    "actix",
-    "axum",
-  ],
-  [
-    "actix",
-    "generic",
-  ],
-  [
-    "generic",
-    "axum",
-  ],
-  [
-    "nightly",
-  ],
-]
-max_combination_size = 2
-
-[package.metadata.docs.rs]
-rustdoc-args = ["--generate-link-to-definition"]
-
-[lints.rust]
-unexpected_cfgs = { level = "warn", check-cfg = ['cfg(rustc_nightly)'] }
+__internal_erase_components = ["leptos_macro_core/__internal_erase_components"]
