+++ leptos_macro/leptos_macro_core/Cargo.toml
@@ -1,0 +1,99 @@
+[package]
+name = "leptos_macro_core"
+version = { workspace = true }
+authors = ["Greg Johnston"]
+license = "MIT"
+repository = "https://github.com/leptos-rs/leptos"
+description = "Internal code for the view macro for the Leptos web framework."
+readme = "../../README.md"
+rust-version.workspace = true
+edition.workspace = true
+
+[dependencies]
+attribute-derive = { version = "0.10.3", features = ["syn-full"] }
+cfg-if = "1.0"
+html-escape = "0.2.13"
+itertools = { workspace = true }
+prettyplease = "0.2.25"
+proc-macro-error2 = { version = "2.0", default-features = false }
+proc-macro2 = "1.0"
+quote = "1.0"
+syn = { version = "2.0", features = ["full"] }
+rstml = "0.12.0"
+leptos_hot_reload = { workspace = true }
+server_fn_macro = { workspace = true }
+convert_case = { workspace = true }
+uuid = { version = "1.11", features = ["v4"] }
+tracing = { version = "0.1.41", optional = true }
+
+[dev-dependencies]
+log = "0.4.22"
+typed-builder = "0.20.0"
+trybuild = { workspace = true }
+leptos = { workspace = true }
+leptos_router = { workspace = true, features = ["ssr"] }
+server_fn = { workspace = true, features = ["cbor"] }
+insta = "1.41"
+serde = "1.0"
+
+[build-dependencies]
+rustc_version = "0.4.1"
+
+[features]
+csr = []
+hydrate = []
+ssr = ["server_fn_macro/ssr", "leptos/ssr"]
+nightly = ["server_fn_macro/nightly"]
+tracing = ["dep:tracing"]
+islands = []
+trace-components = []
+trace-component-props = []
+actix = ["server_fn_macro/actix"]
+axum = ["server_fn_macro/axum"]
+generic = ["server_fn_macro/generic"]
+# Having an erasure feature rather than normal --cfg erase_components for the proc macro crate is a workaround for this rust issue:
+# https://github.com/rust-lang/cargo/issues/4423
+# TLDR proc macros will ignore RUSTFLAGS when --target is specified on the cargo command.
+# This works around the issue by the non proc-macro crate which does see RUSTFLAGS enabling the replacement feature on the proc-macro crate, which wouldn't.
+# This is automatic as long as the leptos crate is depended upon,
+# downstream usage should never manually enable this feature.
+__internal_erase_components = []
+
+[package.metadata.cargo-all-features]
+denylist = ["tracing", "trace-component-props", "trace-components"]
+skip_feature_sets = [
+    [
+        "csr",
+        "hydrate",
+    ],
+    [
+        "hydrate",
+        "csr",
+    ],
+    [
+        "hydrate",
+        "ssr",
+    ],
+    [
+        "actix",
+        "axum",
+    ],
+    [
+        "actix",
+        "generic",
+    ],
+    [
+        "generic",
+        "axum",
+    ],
+    [
+        "nightly",
+    ],
+]
+max_combination_size = 2
+
+[package.metadata.docs.rs]
+rustdoc-args = ["--generate-link-to-definition"]
+
+[lints.rust]
+unexpected_cfgs = { level = "warn", check-cfg = ['cfg(rustc_nightly)'] }
