+++ leptos_macro/leptos_macro_core/src/lazy.rs
